+++ smol/moca/processor/ewald.py
@@ -20,7 +20,7 @@ from pymatgen.analysis.ewald import EwaldSummation
 from smol.cofe.extern.ewald import EwaldTerm
 from smol.cofe.space.clusterspace import ClusterSubspace
 from smol.moca.processor.base import Processor
-from smol.utils.cluster_utils.ewald import delta_ewald_single_flip
+from smol.utils.cluster.ewald import delta_ewald_single_flip
 
 
 class EwaldProcessor(Processor):
