+++ smol/utils/cluster/struct.pxd
